Refactor template dialog and chat.update apis
This code can be shared by the project creation handlers.
